requirements.txt
altair==4.1.0
astor==0.8.1
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.4
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.12.41
botocore==1.15.41
breadability==0.1.20
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gensim==3.8.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.2.4
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.34
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.1.1
plac==1.1.3
plotly==4.4.1
plotly-express==0.4.0
pluggy==0.13.1
preprocessing==0.1.13
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
py==1.8.1
pycountry==19.8.18
pydeck==0.3.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-timeit==0.3.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
regex==2020.4.4
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.13.0
smart-open==1.11.1
spacy==2.2.4
sphinx-rtd-theme==0.2.4
srsly==1.0.2
statsmodels==0.10.2
streamlit==0.57.3
sumy==0.8.1
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.45.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
validators==0.14.3
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
